make plugin more discoverable
