+++ .github/workflows/test.yaml
@@ -128,7 +128,7 @@ jobs:
           DD_SERVICE: weave-python
           DD_ENV: ci
           # PARQUET_ENABLED: true # TODO: Enable parquet after available in CI
-        run: WB_SERVER_HOST=http://wandbservice WEAVE_SERVER_DISABLE_ECOSYSTEM=1 source /root/venv/bin/activate && cd weave && pip install -e . && pytest -p weave.pytest_plugins.pytest_report_time ./tests ./ops_arrow --ddtrace
+        run: WB_SERVER_HOST=http://wandbservice WEAVE_SERVER_DISABLE_ECOSYSTEM=1 source /root/venv/bin/activate && cd weave && pytest ./tests ./ops_arrow --ddtrace
 
   ecosystem-tests:
     name: Python ecosystem unit tests
@@ -147,7 +147,7 @@ jobs:
         env:
           DD_SERVICE: weave-python
           DD_ENV: ci
-        run: source /root/venv/bin/activate && cd weave && pip install -e . && pytest -p weave.pytest_plugins.pytest_report_time ./ecosystem --ddtrace
+        run: source /root/venv/bin/activate && cd weave && pytest ./ecosystem --ddtrace
 
   # nbmake:
   #   name: Run notebooks with nbmake
